Frequencies generated are no longer integer Hz

+++ samplegenerator.hpp
@@ -12,7 +12,7 @@ public:
 
     int wave_shape;
 
-    int toHz();
+    double toHz();
     void fillSamples(double *, int, double, int);
 };
